Modified - move store provider
