+++ lib/features/customer_feature/presentation/screens/login_screen.dart
@@ -5,7 +5,7 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:provider/provider.dart';
 
 import '../../../../core/connection/http_exception.dart';
-import '../../../../provider/Products.dart';
+import '../../../store_feature/presentation/providers/Products.dart';
 import '../../../../core/theme/app_theme.dart';
 import '../providers/auth.dart';
 import '../../../../core/widgets/main_drawer.dart';
