+++ lib/features/store_feature/presentation/providers/Products.dart
@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../models/category.dart';
-import '../models/color_code_card.dart';
-import '../models/color_code_product_detail.dart';
-import '../models/order_send_details.dart';
-import '../models/product.dart';
-import '../models/product_cart.dart';
-import '../models/product_main.dart';
-import '../models/search_detail.dart';
-import '../core/constants/urls.dart';
+import '../../../../models/category.dart';
+import '../../../../models/color_code_card.dart';
+import '../../../../models/color_code_product_detail.dart';
+import '../../../../models/order_send_details.dart';
+import '../../../../models/product.dart';
+import '../../../../models/product_cart.dart';
+import '../../../../models/product_main.dart';
+import '../../../../models/search_detail.dart';
+import '../../../../core/constants/urls.dart';
 
 class Products with ChangeNotifier {
   List<Product> _items = [];
