+++ lib/features/store_feature/presentation/screens/product_detail_screen.dart
@@ -8,7 +8,7 @@ import 'package:provider/provider.dart';
 import 'package:tamizshahr/core/widgets/buton_bottom.dart';
 
 import '../../../../models/product.dart';
-import '../../../../provider/Products.dart';
+import '../providers/Products.dart';
 import '../../../../core/theme/app_theme.dart';
 import '../../../../core/logic/en_to_ar_number_convertor.dart';
 import '../../../../core/widgets/main_drawer.dart';
