+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:tamizshahr/provider/charities.dart';
 import 'package:tamizshahr/provider/clearings.dart';
-import 'package:tamizshahr/provider/orders.dart';
+import 'package:tamizshahr/features/store_feature/presentation/providers/orders.dart';
 import 'package:tamizshahr/screens/charity_detail_screen.dart';
 import 'package:tamizshahr/screens/charity_screen.dart';
 import 'package:tamizshahr/screens/clear_screen.dart';
@@ -36,7 +36,7 @@ import 'features/waste_feature/presentation/waste_request_date_screen.dart';
 import 'features/waste_feature/presentation/waste_request_send_screen.dart';
 import 'features/waste_feature/presentation/wastes_screen.dart';
 import 'core/constants/strings.dart';
-import 'provider/Products.dart';
+import 'features/store_feature/presentation/providers/Products.dart';
 import 'features/customer_feature/presentation/providers/customer_info.dart';
 import 'features/waste_feature/collect_list_screen.dart';
 import 'features/customer_feature/presentation/screens/customer_detail_info_edit_screen.dart';
