+++ lib/widgets/card_item.dart
@@ -4,7 +4,7 @@ import 'package:intl/intl.dart' as intl;
 import 'package:provider/provider.dart';
 
 import '../models/product_cart.dart';
-import '../provider/Products.dart';
+import '../features/store_feature/presentation/providers/Products.dart';
 import '../core/theme/app_theme.dart';
 import '../features/customer_feature/presentation/providers/auth.dart';
 import '../features/store_feature/presentation/screens/product_detail_screen.dart';
